Separate glyph and label components
