+++ packages/svg/src/SvgFeatureRenderer/components/SvgFeatureRendering.js
@@ -1,10 +1,102 @@
+import { readConfObject } from '@gmod/jbrowse-core/configuration'
 import { PropTypes as CommonPropTypes } from '@gmod/jbrowse-core/mst-types'
+import { featureSpanPx } from '@gmod/jbrowse-core/util'
+import SceneGraph from '@gmod/jbrowse-core/util/layouts/SceneGraph'
 import { observer } from 'mobx-react'
 import ReactPropTypes from 'prop-types'
 import React from 'react'
 import Box from './Box'
 import Chevron from './Chevron'
-import './SvgFeatureRendering.scss'
+import FeatureLabel from './FeatureLabel'
+
+const fontWidthScaleFactor = 0.55
+
+function layOut(args) {
+  const { feature, bpPerPx, region, layout, horizontallyFlipped } = args
+
+  const [startPx, endPx] = featureSpanPx(
+    feature,
+    region,
+    bpPerPx,
+    horizontallyFlipped,
+  )
+  const rootLayout = new SceneGraph('root', startPx, 0, 0, 0)
+  const featureHeight = readConfObject(args.config, 'height', [feature])
+  const featureWidth = endPx - startPx
+  rootLayout.addChild('feature', 0, 0, featureWidth, featureHeight)
+
+  const name = readConfObject(args.config, ['labels', 'name'], [feature]) || ''
+  const description =
+    readConfObject(args.config, ['labels', 'description'], [feature]) || ''
+  const fontHeight = readConfObject(
+    args.config,
+    ['labels', 'fontSize'],
+    ['feature'],
+  )
+  const fontWidth = fontHeight * fontWidthScaleFactor
+  const shouldShowName = /\S/.test(name)
+  const shouldShowDescription = /\S/.test(description)
+  const textVerticalPadding = 2
+  let labelWidth
+  let descriptionWidth
+  const maxFeatureGlyphExpansion = readConfObject(
+    args.config,
+    'maxFeatureGlyphExpansion',
+  )
+  if (shouldShowName) {
+    labelWidth = Math.round(
+      Math.min(
+        name.length * fontWidth,
+        featureWidth + maxFeatureGlyphExpansion,
+      ),
+    )
+
+    rootLayout.addChild(
+      'nameLabel',
+      0,
+      rootLayout.getSubRecord('feature').bottom + textVerticalPadding,
+      labelWidth,
+      fontHeight,
+    )
+  }
+  if (shouldShowDescription) {
+    descriptionWidth = Math.round(
+      Math.min(
+        description.length * fontWidth,
+        featureWidth + maxFeatureGlyphExpansion,
+      ),
+    )
+    rootLayout.addChild(
+      'descriptionLabel',
+      0,
+      rootLayout.getSubRecord(shouldShowName ? 'nameLabel' : 'feature').bottom +
+        textVerticalPadding,
+      descriptionWidth,
+      fontHeight,
+    )
+  }
+
+  const start = feature.get('start')
+  const topPx = layout.addRect(
+    feature.id(),
+    start,
+    start + rootLayout.width * bpPerPx,
+    rootLayout.height,
+  )
+
+  rootLayout.move(0, topPx)
+
+  return {
+    rootLayout,
+    name,
+    description,
+    shouldShowDescription,
+    shouldShowName,
+    fontHeight,
+    labelWidth,
+    descriptionWidth,
+  }
+}
 
 function SvgFeatureRendering(props) {
   const {
@@ -67,8 +159,8 @@ function SvgFeatureRendering(props) {
 
   const featuresRendered = []
   for (const feature of features.values()) {
-    const FeatureComponent = chooseGlyphComponent(feature)
-    const layoutRecord = FeatureComponent.layout({
+    const GlyphComponent = chooseGlyphComponent(feature)
+    const layoutRecord = (GlyphComponent.layOut || layOut)({
       feature,
       horizontallyFlipped,
       bpPerPx,
@@ -76,14 +168,55 @@ function SvgFeatureRendering(props) {
       config,
       layout,
     })
+    const {
+      rootLayout,
+      name,
+      description,
+      shouldShowDescription,
+      shouldShowName,
+      fontHeight,
+    } = layoutRecord
+    const exp = readConfObject(config, 'maxFeatureGlyphExpansion') || 0
+
+    const featureLayout = rootLayout.getSubRecord('feature')
     featuresRendered.push(
-      <FeatureComponent
-        {...props}
-        layoutRecord={layoutRecord}
-        feature={feature}
+      <g
+        transform={`translate(${rootLayout.left} ${rootLayout.top})`}
         key={feature.id()}
-        selectedFeatureId={selectedFeatureId}
-      />,
+      >
+        <GlyphComponent
+          {...props}
+          feature={feature}
+          featureLayout={featureLayout}
+          selected={String(selectedFeatureId) === String(feature.id())}
+        />
+        {!shouldShowName ? null : (
+          <FeatureLabel
+            text={name}
+            x={rootLayout.getSubRecord('nameLabel').left}
+            y={rootLayout.getSubRecord('nameLabel').top}
+            color={readConfObject(config, ['labels', 'nameColor'], [feature])}
+            fontHeight={fontHeight}
+            featureWidth={featureLayout.width}
+            allowedWidthExpansion={exp}
+          />
+        )}
+        {!shouldShowDescription ? null : (
+          <FeatureLabel
+            text={description}
+            x={rootLayout.getSubRecord('descriptionLabel').left}
+            y={rootLayout.getSubRecord('descriptionLabel').top}
+            color={readConfObject(
+              config,
+              ['labels', 'descriptionColor'],
+              [feature],
+            )}
+            fontHeight={fontHeight}
+            featureWidth={featureLayout.width}
+            allowedWidthExpansion={exp}
+          />
+        )}
+      </g>,
     )
   }
