--- packages/svg/src/SvgFeatureRenderer/components/SvgFeatureRendering.scss
@@ -1,13 +1,0 @@
-.PileupRendering {
-  color: black;
-
-    white-space: normal;
-
-    .feature {
-    display: inline-block;
-    height: 1.2em;
-    margin: 1px;
-    background: blue;
-    color: white;
-  }
-}
